ipc: tell VM to watch a process once for any IPC call
When the interprocess communication server receives an IPC system
call from a process, ipc should tell VM to watch it only one. This
patch fixes that. In the previous code, ipc was telling VM to watch
the process for each iteration of the matching loop. In the case
where the call type is matched in the last iteration, ipc would
tell VM to watch the process numerous times.

The patch was tested on Minix 10.0.2.15 3.2.1 i686.

Signed-off-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>