ipc: use an offset in the system call table for faster handling
Instead of using a loop to find a matching ipc (inter process
communication) system call type, the offset in the call table can be
simply calculated in constant time.

Also, when the interprocess communication server receives an ipc
system call from a process, ipc should tell VM to watch it only one.
This patch fixes that also. In the previous code, ipc was telling VM
to watch the process for each iteration of the matching loop. In the
case where the call type is matched in the last iteration, ipc would
tell VM to watch the process numerous times.

This patch was tested on Minix 10.0.2.15 3.2.1 i686 inside VirtualBox.

An initial version (in 2 separate patches) of this patch was reviewed
by David C. van Moolenbroek.

Signed-off-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>